Added Akari voice lines, minor fixes
