+++ package.json
@@ -4,7 +4,7 @@
   "description": "Puyo Puyo for your browser.",
   "main": "Main.js",
   "scripts": {
-    "bundle": "./node_modules/.bin/watchify js/main.js -o bundle.js --debug",
+    "bundle": "./node_modules/.bin/watchify js/Main.js -o bundle.js --debug",
     "start": "node js/server.js"
   },
   "repository": {
