+++ sounds/voices/chains/chain_10_jpn.wav
