+++ sounds/voices/chains/chain_11_jpn.wav
