+++ sounds/voices/chains/chain_12_jpn.wav
