+++ sounds/voices/chains/chain_1_jpn.wav
