+++ sounds/voices/chains/chain_2_jpn.wav
