+++ sounds/voices/chains/chain_3_jpn.wav
