+++ sounds/voices/chains/chain_4_jpn.wav
