+++ sounds/voices/chains/chain_5_jpn.wav
