+++ sounds/voices/chains/chain_6_jpn.wav
