+++ sounds/voices/chains/chain_7_jpn.wav
