+++ sounds/voices/chains/chain_8_jpn.wav
