+++ sounds/voices/chains/chain_9_jpn.wav
